+++ go.js
@@ -16,36 +16,36 @@ stop('analytics')
 await new Promise(resolve => setTimeout(resolve, 500))
 stop('database')
 
-console.log('TIMER COUNT 3', timers.timersList().length)
-assert.equal(timers.timersList().length, 3)
+console.log('TIMER COUNT 3', timers.timers().length)
+assert.equal(timers.timers().length, 3)
 
-console.log(timers.headerKey)
-assert.equal(timers.headerKey, 'Server-Timing')
+console.log(timers.key)
+assert.equal(timers.key, 'Server-Timing')
 
 console.log(timers.values())
-console.log(timers.headerValue())
-console.log(timers.headerObject())
-console.log(timers.fullHeaderString())
+console.log(timers.value())
+console.log(timers.toObject())
+console.log(timers.toString())
 
 start("this won't work")
 timers.reset()
 stop("this won't work")
 
-console.log('TIMER COUNT 0', timers.timersList().length)
-assert.equal(timers.timersList().length, 0)
+console.log('TIMER COUNT 0', timers.timers().length)
+assert.equal(timers.timers().length, 0)
 
 start('reset')
 await new Promise(resolve => setTimeout(resolve, 100))
 stop('reset')
 
-console.log('TIMER COUNT 1', timers.timersList().length)
-assert.equal(timers.timersList().length, 1)
+console.log('TIMER COUNT 1', timers.timers().length)
+assert.equal(timers.timers().length, 1)
 
 console.log(timers.values())
 
 timers.reset()
-console.log('TIMER COUNT 0', timers.timersList().length)
-assert.equal(timers.timersList().length, 0)
+console.log('TIMER COUNT 0', timers.timers().length)
+assert.equal(timers.timers().length, 0)
 
 start() // 1
 stop() // stop 1 => <1ms
@@ -56,7 +56,7 @@ start() // 3
 stop() // stop 3 => <1ms
 await new Promise(resolve => setTimeout(resolve, 100))
 stop() // stop 2 => 100+ms
-stop('foo')
+stop('foo') // => 100+ms
 
 console.log(timers.values())
-assert.equal(timers.timersList().length, 4)
+assert.equal(timers.timers().length, 4)